gen: add a set of self-tests
The tests run with a veth pair; the pair is created before running the
tests and destroyed after that.

The tests run ipgen with several options.  Note that the tests just
check if ipgen sends and receives packets without packet drops.  More
detailed tests will be added in the future.