Don't call statistics_clear() in the beginning of RFC2544_RESETTING case.
rfc2544_test() is called every 1 / DISPLAY_UPDATE_HZ second. Currently,
it waits 0.05 second. The transmission is disabled while RFC2544_RESETTING
state, so it's not required to call statistics_clear() function in the
beginning of the case. Moreover, calling statistics_clear()
might hide unexpected packet reception, so it should be removed.